control-plane-api: is_service_account lives with the grant helpers
The predicate is an identity lookup with nothing GraphQL about it, and
grants.rs is the crate's home for identity DB helpers — both its
consumers (the refresh-token GraphQL mutations and the REST
capability_token mint) reach it as crate::grants without one surface
reaching into the other's module. The mint test's copy-through claim
assertions are one redacted insta snapshot of the full serialized
claim set, which pins the payload's exact key set — the Rust half of
the two-snapshot claim parity with the pgTAP-pinned SQL mint.
