+++ crates/control-plane-api/src/server/public/graphql/mod.rs
@@ -40,7 +40,6 @@ mod publication_history;
 mod refresh_tokens;
 mod scalars;
 mod service_accounts;
-pub(crate) use service_accounts::is_service_account;
 pub mod status;
 mod storage_mappings;
 mod tenant;
